% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpm.R
\name{isBadNum}
\alias{isBadNum}
\title{checks if a log-density evaluation is NA or NaN}
\usage{
isBadNum(num)
}
\arguments{
\item{num}{evaluation of a log-density}
}
\value{
TRUE or FALSE
}
\description{
checks if a log-density evaluation is NA or NaN
}
\examples{
isBadNum(NaN)
}
