% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpm.R
\name{mean.cpmResults}
\alias{mean.cpmResults}
\title{calculates the posterior mean point estimate}
\usage{
\method{mean}{cpmResults}(x, ...)
}
\arguments{
\item{x}{a cpmResults object}

\item{...}{arguments to be passed to or from methods.}
}
\value{
a vector of parameter estimates (posterior mean)
}
\description{
calculates the posterior mean point estimate
}
