% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpm.R
\name{plot.cpmResults}
\alias{plot.cpmResults}
\title{plots a cpmResults object}
\usage{
\method{plot}{cpmResults}(x, ...)
}
\arguments{
\item{x}{a cpmResults object}

\item{...}{arguments to be passed to or from methods.}
}
\description{
plots a cpmResults object
}
