% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpm.R
\name{print.cpmResults}
\alias{print.cpmResults}
\title{prints a cpmResults object}
\usage{
\method{print}{cpmResults}(x, ...)
}
\arguments{
\item{x}{a cpmResults object}

\item{...}{arguments to be passed to or from methods.}
}
\value{
the same cpmResults object
}
\description{
prints a cpmResults object
}
